nhandangtengavatrom/requirements.txt
numpy==1.23.5
tensorflow==2.11.0
opencv-python==4.7.0.72
RPi.GPIO==0.7.1
scikit-learn==1.2.2
face-recognition==1.3.0
dlib==19.24.0
imutils==0.5.4
PyYAML==6.0
tqdm==4.65.0
matplotlib==3.7.1
pillow==9.5.0 